+++ Packages/com.cdm.mobiledatatransfer/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-02-20
+- Some Android files moved to AndroidLib.dll.
+- Added handshake to check that the socket is being listened on the Android side.
+- AndroidLib.dll updated.
+
 ## [1.0.5] - 2024-02-09
 - AndroidLib dll changed. (Multiplatform support added.)